Fix issue #7: Correct jQuery reference link
This commit should fix the problem in Issue #7 where enabling the
jQuery section causes JS to crash.  The reason why this wasn't
working was because jQuery actually wasn't being added in
index.html, because the local link was incorrect. (The only file
modified is index.html, because index.js is actually fine.)